Bug 1040432 - Move sdk/system/xul-app to a jsm so that cuddlefish/toolkit loaders can access it easily. r=Mossop
